app.py
import os
import instructor
import asyncio
import json
import glob
from werkzeug.utils import secure_filename
from flask import Flask, render_template, request, redirect, url_for, flash, send_file
from openai import OpenAI
from dotenv import load_dotenv

# --- IMPORTS FROM OUR FILES ---
from models import JobListing
from config import CONFIG, ANALYSIS_PROMPT_TEXT, TAILORING_PROMPT_TEXT

# --- APPLICATION SETUP ---
load_dotenv()
client = instructor.patch(OpenAI(api_key=os.getenv("OPENAI_API_KEY")))
app = Flask(__name__)
# A secret key is needed for flashing messages
app.secret_key = os.urandom(24) 

# --- PLACEHOLDER FUNCTIONS (Will be moved to services in Phase 2) ---
def step_0_create_session_directory(output_base_dir: str, company: str, position: str) -> str:
    """Placeholder - will be implemented in Phase 2"""
    return "placeholder_session"

async def step_1_get_content(source_config: dict, session_path: str) -> str:
    """Placeholder - will be implemented in Phase 2"""
    return "placeholder_path"

def step_2_analyze_job(session_path: str, client: OpenAI, model_name: str) -> str:
    """Placeholder - will be implemented in Phase 2"""
    return "placeholder_path"

def step_3_tailor_resume(session_path: str, base_resume_path: str, client: OpenAI, model_name: str, api_parameters: dict) -> str:
    """Placeholder - will be implemented in Phase 2"""
    return "placeholder_path"

def step_4_assemble_and_create_pdf(session_path: str, base_resume_path: str, pdf_config: dict) -> str:
    """Placeholder - will be implemented in Phase 2"""
    return "placeholder_path"

# --- FLASK ROUTES (EXACT COPY FROM JOBBOT) ---
@app.route('/')
def home():
    """Displays the main form for inputting job details and uploading a resume."""
    return render_template('index.html')

@app.route('/generate', methods=['POST'])
def generate():
    """Handles form submission and starts the resume generation pipeline."""
    job_url = request.form.get('job_url', '').strip()
    job_description = request.form.get('job_description', '').strip()
    resume_file = request.files.get('resume_file')
    
    if not (job_url or job_description):
        flash("Error: Please provide either a job URL or job description.")
        return redirect(url_for('home'))
    
    if not resume_file or not resume_file.filename:
        flash("Error: Please upload a resume file.")
        return redirect(url_for('home'))
    
    # Extract company and position for session naming
    company = request.form.get('company', 'Unknown')
    position = request.form.get('position', 'Unknown')
    
    session_id = step_0_create_session_directory(CONFIG["output_base_dir"], company, position)
    session_path = os.path.join(CONFIG["output_base_dir"], session_id)
    
    # For Phase 1, just show success message
    flash("Success! Basic structure is working. Full pipeline will be implemented in Phase 2.")
    return redirect(url_for('home'))

@app.route('/review/joblisting/<session_id>')
def review_joblisting(session_id):
    """Placeholder for Phase 2"""
    flash("This feature will be implemented in Phase 2.")
    return redirect(url_for('home'))

@app.route('/run/analysis/<session_id>', methods=['POST'])
def run_step2_analysis(session_id):
    """Placeholder for Phase 2"""
    flash("This feature will be implemented in Phase 2.")
    return redirect(url_for('home'))

@app.route('/review/jobanalysis/<session_id>')
def review_jobanalysis(session_id):
    """Placeholder for Phase 2"""
    flash("This feature will be implemented in Phase 2.")
    return redirect(url_for('home'))

@app.route('/run/tailoring/<session_id>', methods=['POST'])
def run_step3_tailoring(session_id):
    """Placeholder for Phase 2"""
    flash("This feature will be implemented in Phase 2.")
    return redirect(url_for('home'))

@app.route('/review/tailoring/<session_id>')
def review_tailoring(session_id):
    """Placeholder for Phase 2"""
    flash("This feature will be implemented in Phase 2.")
    return redirect(url_for('home'))

@app.route('/generate/pdf/<session_id>', methods=['POST'])
def generate_pdf(session_id):
    """Placeholder for Phase 2"""
    flash("This feature will be implemented in Phase 2.")
    return redirect(url_for('home'))

@app.route('/review/pdf/<session_id>')
def review_pdf(session_id):
    """Placeholder for Phase 2"""
    flash("This feature will be implemented in Phase 2.")
    return redirect(url_for('home'))

@app.route('/download/pdf/<session_id>')
def download_pdf(session_id):
    """Placeholder for Phase 2"""
    flash("This feature will be implemented in Phase 2.")
    return redirect(url_for('home'))

# --- MAIN EXECUTION ---
if __name__ == '__main__':
    app.run(debug=True)
