requirements.txt
crawl4ai==0.7.2
Flask==3.1.2
instructor==1.11.2
Jinja2==3.1.6
openai==1.106.1
pydantic==2.11.7
python-dotenv==1.1.1
weasyprint==66.0
Werkzeug==3.1.3
PyYAML==6.0.2